refactor: ゲームのコア部分を kero3 パッケージに切り出すように
